Add new implementation of initial intcode computer
The earlier implementation was proving difficult to extend for the day
09 puzzle, so it's been changed from an inheritance based class that
passes around getters and setters to an object composition based class.

Now the operations have direct access to the intcode computer's state,
reducing complexity and increasing readability of the operations.

In addition to this change, the software will be treated as character
strings instead of integers. This is in preperation for splitting out
the parameter modes, and handling arbitrarily large integers.